Remove all implementation from mempool.go
